Refactor Editor Component: Remove unused block selection logic and integrate initial selection restoration
- Removed the `useEffect` hook for restoring block selection from the editor component.
- Integrated initial selection restoration logic into the `ExperimentalEditorProvider` for better handling of navigation scenarios.
- Cleaned up imports by removing unused dependencies related to block selection.
