+++ packages/editor/src/components/editor/index.js
@@ -1,19 +1,16 @@
 /**
  * WordPress dependencies
  */
-import { useSelect, useDispatch } from '@wordpress/data';
+import { useSelect } from '@wordpress/data';
 import { store as coreStore } from '@wordpress/core-data';
 import { Notice } from '@wordpress/components';
 import { __ } from '@wordpress/i18n';
-import { useEffect } from '@wordpress/element';
-import { store as blockEditorStore } from '@wordpress/block-editor';
 
 /**
  * Internal dependencies
  */
 import { store as editorStore } from '../../store';
 import { TEMPLATE_POST_TYPE } from '../../store/constants';
-import { useRestoreBlockFromPath } from '../../utils/block-selection-path';
 import EditorInterface from '../editor-interface';
 import { ExperimentalEditorProvider } from '../provider';
 import Sidebar from '../sidebar';
@@ -98,32 +95,6 @@ function Editor( {
 		[ postType, postId, templateId ]
 	);
 
-	const { selectBlock } = useDispatch( blockEditorStore );
-	const restoreBlockFromPath = useRestoreBlockFromPath();
-
-	// Restore initial block selection if provided (e.g., from navigation)
-	useEffect( () => {
-		if ( ! initialSelection || ! hasLoadedPost || ! post ) {
-			return;
-		}
-
-		// Use setTimeout to ensure blocks are fully rendered before selecting
-		const timeoutId = setTimeout( () => {
-			const clientId = restoreBlockFromPath( initialSelection );
-			if ( clientId ) {
-				selectBlock( clientId );
-			}
-		}, 0 );
-
-		return () => clearTimeout( timeoutId );
-	}, [
-		initialSelection,
-		hasLoadedPost,
-		post,
-		selectBlock,
-		restoreBlockFromPath,
-	] );
-
 	return (
 		<>
 			{ hasLoadedPost && ! post && (
@@ -144,6 +115,7 @@ function Editor( {
 					__unstableTemplate={ template }
 					settings={ settings }
 					initialEdits={ initialEdits }
+					initialSelection={ initialSelection }
 					useSubRegistry={ false }
 				>
 					<EditorInterface { ...props }>
